RAG/YT Chatbot/requirements.txt
streamlit
youtube-transcript-api
langchain-community
langchain-openai
faiss-cpu
tiktoken
python-dotenv
langchain-huggingface
langchain-groq
sentence-transformers